test: only test postgres if POSTGRES_URL is set

+++ tests/helpers/server.mjs
@@ -5,7 +5,7 @@ import { prepareHttpServer } from "../../src/http.mjs";
 import { prepareDatabase } from "../../src/database.mjs";
 import postgres from "postgres";
 import { setSchema } from "@konsumation/db-postgresql";
-import getPort from '@ava/get-port';
+import getPort from "@ava/get-port";
 
 function pn(path) {
   return fileURLToPath(new URL(path, import.meta.url));
@@ -119,10 +119,9 @@ export async function wait(msecs = 1000) {
   await new Promise(resolve => setTimeout(resolve, msecs));
 }
 
-
 export async function* allContexts(context, users, dataFile) {
-  for (const db of [
-    async (port) => {
+  const dbs = [
+    async port => {
       const databaseFile = pn(`../build/db-${port}`);
       return {
         database: { "@konsumation/db-level": databaseFile },
@@ -130,8 +129,11 @@ export async function* allContexts(context, users, dataFile) {
           await rm(databaseFile, { recursive: true });
         }
       };
-    },
-    async (port) => {
+    }
+  ];
+
+  if (process.env.POSTGRES_URL) {
+    dbs.push(async port => {
       const schemaName = "testintegration";
       const sql = postgres(process.env.POSTGRES_URL);
       await sql`DROP SCHEMA IF EXISTS ${sql(schemaName)} CASCADE`;
@@ -148,8 +150,10 @@ export async function* allContexts(context, users, dataFile) {
           await sql.end();
         }
       };
-    }
-  ]) {
+    });
+  }
+
+  for (const db of dbs) {
     const port = await getPort();
     const { database, cleanup } = await db(port);
     await startServer(context, port, database, users, dataFile);
